refactor: 게시글/댓글 도메인 구조 개선 및 작성자 검증 추가

+++ minisns/src/main/java/com/example/minisns/service/CommentService.java
@@ -10,6 +10,7 @@ import com.example.minisns.exception.UserNotFoundException;
 import com.example.minisns.repository.CommentRepository;
 import com.example.minisns.repository.PostRepository;
 import com.example.minisns.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,21 +20,17 @@ import java.util.List;
 
 @Service
 @Transactional
+@RequiredArgsConstructor
 public class CommentService {
+
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
-    public CommentService(CommentRepository commentRepository, PostRepository postRepository, UserRepository userRepository) {
-        this.commentRepository = commentRepository;
-        this.postRepository = postRepository;
-        this.userRepository = userRepository;
-    }
-
     public CommentResponse create(Long postId, Long userId, String content) {
-        Post post = findByPostId(postId);
-        User user = findByUserId(userId);
-        Comment comment = new Comment(post, user, content);
+        Post post = findPostById(postId);
+        User user = findUserById(userId);
+        Comment comment = Comment.create(post, user, content);
         return CommentResponse.toResponse(commentRepository.save(comment));
     }
 
@@ -45,24 +42,32 @@ public class CommentService {
         return commentRepository.findResponsesByPostId(postId, pageable); // Page<Comment>반환
     }
 
-    public void delete(Long id) {
-        Comment comment = findByCommentId(id);
+    public void delete(Long loginUserId, Long commentId) {
+        Comment comment = findCommentById(commentId);
+        validateCommentOwner(loginUserId, comment);
         commentRepository.delete(comment);
     }
 
-    private Post findByPostId(Long postId) {
+    // =============== PRIVATE METHOD =================
+
+    private Post findPostById(Long postId) {
         return postRepository.findById(postId)
                 .orElseThrow(() -> new PostNotFoundException(postId));
     }
 
-    private User findByUserId(Long userId) {
+    private User findUserById(Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(userId));
     }
 
-    private Comment findByCommentId(Long id) {
+    private Comment findCommentById(Long id) {
         return commentRepository.findById(id)
                 .orElseThrow(() -> new CommentNotFoundException(id));
     }
 
+    private void validateCommentOwner(Long loginUserId, Comment comment) {
+        if (!comment.isOwner(loginUserId)) {
+            throw new IllegalArgumentException("댓글 작성자만 삭제할 수 있습니다.");
+        }
+    }
 }
